Make dead keys behave as expected when followed by certain keys (under Windows).
Since 7.4.413 dead keys' nominal characters were no longer output as expected when typing numbers with the num pad or hitting return after hitting a dead key (for example).
This was because those keys are dealt with independently and did not consume the dead key as a Windows user would expect.