Tidy-up an error check in vim_findfile_init().
The function contains a memory allocaton error check
with regard to wildcard handling, but this code fragment
path is triggered even if the path to search contains no wildcards.
This causes vim_findfile_init() to fail on a non-existent start
directory.

Only perform the error handling if the search path contains
wildcards.

This makes upwards search work if the start directory constructed
by vim_findfile_init does not exist.