Remove clip_init(FALSE) call when 'term' is set
Using a Vim that is built to be able to access X's PRIMARY and CLIPBOARD
selections (i.e., "* and "+), performing “:let &term=&term” causes Vim
to "forget" that it knows how to access those registers.

This is because set_termname() ends up calling clip_init(), which sets
owned = FALSE on both VimClipboard instances.

Changing &term should not modify the state of the clipboard at all, so this
patch removes that bit of the code, resolving the bug.