Support 4608 colour terminals.
Vim assumed that all terminals that do not have 8, 16, 88 or 256 colours
have a particular non-ansi, non-terminfo ordering to the first 16 colours.
Terminals that had one of those counts would only get ANSI ordering if
the terminfo control sequence ends with a letter "m".

I have left the weirdness in, but added the assumption that a terminal
with more than 256 colours will do the sane thing and use the XTerm-256
ordering for these colours.

If it does not, terminfo is sufficiently programmable to make it happen.