Rebuild the autoconf files during CI testing
Although the autoconf related files don't change often, it's useful to
ensure they can be rebuilt.  This will help catch possible breakage and
make it easier for pull requests which modify src/configure.in to get
sanity tested.

In order to ensure the checked in configure also works, the autoconf
files are only rebuilt when $CHECK_AUTOCONF=yes.