matchparen: Show matches of multi-byte chars
Using an index into the buffer’s line only gives a single byte of a
multi-byte character.  When 'matchpairs' contains multi-byte characters
(e.g., :set matchpairs+=（:）,［:］,【:】,『:』,「:」), the matching
pair isn't properly highlighted.

Using matchlist() to get the character at the byte offset of the
cursor, as well as the preceeding character, enables showing the match
of a multi-byte pair.