Fix E315 when switching buffers
When switching to a new buffer, the CursorMoved autocommand is triggered
for the new buffer. If there is a vim script that applies syntax
highlighting in that autocommand, syntax_start will be called while
syn_buf has not been updated to the new buffer. This will cause "E315
ml_get: invalid lnum" as syn_buf is still pointing to the old buffer but
lnum is taken from the new buffer. The additional check for syn_buf !=
wp_buffer causes an additional invalidate_current_state() and sets
syn_buf to the correct buffer.