NUL-terminate *fnamep before passing it to vim_strsave_shellescape.
Some of the transformations that occur earlier in modify_fname simply
adjust *fnamelen to indicate to the caller what portion of *fnamep to
use as the modified filename.  Since vim_strsave_shellescape expects a
NUL-terminated string (so it can duplicate it), ensure *fnamep has been
terminated appropriately before the call.

This fixes issue #706.