Avoid highlighting the new matrix multiplication operator.
The symbol for matrix multiplication (introduced in Python 3.5) is the
same as the symbol for the start of the decorator. This change should
correctly highlight only the decorators, but avoid highlighting the
multiplication operator. It also avoids highlighting syntactically
incorrect use of a function definition on the same line with a
decorator (or any other code).