Add bvimgrep command
This patch adds the :bvimgrep and :bvimgrepadd command.
This is allows to vimgrep for all available buffers and has the
advantage, that vim does not need to load files from disk.

Also this allows to execute :vimgrep in buffers, that have no
corresponding file on disk (or which are changed for which :vimgrep
would only run on the disk-version).

Usage:
:vimgrep[!] <buffer> /{pattern}/[g][j] [{buffer} ...]
:vimgrep[!] <buffer> {pattern} [{buffer} ...]
			Same as ":vimgrep", except that it searches the given
			buffers. If no buffer is given, searches the current
			buffer. Buffer can be given as buffer name (including
			file-patterns) or number.

						*:bvimgrepa* *:bvimgrepadd*
:vimgrepa[dd][!] <buffer> /{pattern}/[g][j] [{buffer} ...]
:vimgrepa[dd][!] <buffer> {pattern} [{buffer} ...]
			Just like ":bvimgrep", but instead of making a new
			list of errors thep matches are appended to the
			current list.

Changes:
20170920:
-Applies cleanly to current HEAD (v8.0.1128)
20170328:
-Use <buffer> argument to :vimgrep command instead of separate :bvimgrep
 commands
20161108:
- Add :blvimgrep and :blvimgrepadd command
201606:
- Add tests for :blvimgrep[add] commands
- initial version