Fix builds against non-Framework Pythons on OS X
Python may optionally build in the usual Unix layout or in the OS
X-style Framework layout on OS X. The configure script was assuming that
Python was always a framework. Apple's system python is a framework
build, so there is always a framework available, so it was being
referenced even when the desired Python was not a framework. When the
system and desired Pythons were built with incompatible options, this
led to build failures.