Fix: match for matchaddpos with negative priority incorrectly highlighted as hlsearch.
The following demonstrates the problem:
    set hlsearch
    let @/=''
    0put='---'
    call matchaddpos('ErrorMsg', [1])
    1put='*** This should be highlighted as ErrorMsg (as the line above), not as Search. ***'
    call matchaddpos('ErrorMsg', [2], -1)

About the fix:
The next_search_hl function sets fields in a match_T. The function can use a matchitem_T* which is passed as its last argument. There is no matchitem_T for the hlsearch in which case I think NULL should be (but wasn't) passed as the last argument to next_search_hl.

(An alternate solution would be to have next_search_hl check if the matchitem_T* passed corresponds with the match_T* passed. However I think the code is probably easier to understand when NULL is passed when there is no matchitem_T corresponding to the match_T* passed.)