find_tags: Use a hashtab_T instead of garray_T to collect tag results
As discussed in #1044, much of the time currently spent in finding tags
is looking for duplicate tags.  Since the tags are stored in a garray_T,
this has to perform a linear search of all existing tags every time a
new tag is added.

Using a hashtab_T (with a modified hash function to allow embedded NUL
bytes), the deduplication costs are drastically cheaper.

Comparing the times to find all tags containing the letter a in a tags
file for all of /usr/include:

  $ time ./src/vim.hash -u NONE --cmd 'echo len(taglist("a"))' --cmd q
  115305
  ./src/vim -u NONE --cmd 'echo len(taglist("a"))' --cmd q  0.69s user 0.04s system 98% cpu 0.741 total
  $ time ./src/vim -u NONE --cmd 'echo len(taglist("a"))' --cmd q
  115305
  ./src/vim -u NONE --cmd 'echo len(taglist("a"))' --cmd q  80.69s user 0.08s system 99% cpu 1:20.90 total