Fix widget sizing on current Gtk+ 3.22 branch
Gtk recently started emitting configure events when repositioning child
windows[1]. Events for the form's child windows bubble up the widget
hierarchy, get caught by our configure event handler, and treated as if
they were reporting the size of the form window. As a result, the shell
was sized using far too small dimensions[2][3]. This change will be part
of Gtk+ 3.22.2.

Solve this by checking whether the configure event we get is really for
the form widget.

[1]: https://git.gnome.org/browse/gtk+/commit/?h=gtk-3-22&id=12579fe71b3b8f79eb9c1b80e429443bcc437dd0
[2]: https://bugs.archlinux.org/task/51509
[3]: https://bugzilla.gnome.org/show_bug.cgi?id=773387