Handling of CamelCase in the spell checker
The problem is that my source code has a lot of camel case identifiers,
dictated by our coding standard. I want to use the spell checker but
it marks every camel case identifier as a miss-spelled word.

This is a simple patch to the spell checking that interprets a lower
case letter followed by an uppercase as a word boundary.

Change-Id: I6f581d0bd15f2a41787df8db6853bd4d3db92913