Fix parallel make of object files
The objects target creates the objects directory.  The object files are
created inside the objects directory, but the object file targets don't
specify the objects target as a dependency.  For a parallel make (e.g.,
make -j 4), this can lead to a build failure if an object file target is
executed before the objects target is executed.

For example, if the objects/os_macosx.o target is executed before the
objects target, then the build will try to create os_macosx.o inside the
objects directory, but that will fail because the objects directory does
not exist yet.

To fix this, add the objects target as a dependency of all object file
targets, and add a -p option to mkdir in the objects target since the
objects directory can already exist.