Make object file targets create objects directory
To support parallel make by ensuring that the objects directory gets
created before the object file targets write to it, the objects target
was added as a dependency of the object file targets.  This had the
unfortunate side effect of causing all object files to be built every
time (I assume because writing to the objects directory caused the
last-modified time of the objects directory to be updated).  Revert that
dependency, and instead make each object file target create the objects
directory before creating the object file in it.