Make :stag consider the 'switchbuf' option
The useopen and usetab values of 'switchbuf' should (according to :help
'swb') be _used in all buffer related split commands, for example
":sbuffer", ":sbnext", or ":sbrewind"._

The :stag / :stselect / :stjump commands also split windows and open a
new buffer.  Therefore, users expect them to be influenced by
'switchbuf' as well, but they currently are not.  When navigating tags,
:set switchbuf=useopen can have great benefits: It avoids that multiple
buffers with the same source file are opened; users will end up with the
minimum amount of split windows, just one for each source file, and do
not have to manually manage duplicate window splits.

The implementation in jumpto_tag() parallels the one in
buflist_getfile(). getfile_result, already used later as a flag is now
initialized at the beginning of the function, and used to actually skip
the getfile() call if a window reuse has happened.