Do `BufEnter` for directories.
Abuse NOTDONE to give some nuance to the return value of readfile(), so
that open_buffer() can distinguish between "general failure" and "failed
because the path is a directory".

Before this change, Vim *already* creates a new buffer when a directory
is edited. So there's no reason it should not raise BufEnter, that was
an implementation detail of readfile().

Most of the changes in this commit merely preserve the old semantics.
The "implicit" change that we actually are interested in, is this line
in `open_buffer()`, where `&retval` being non-FAIL allows EVENT_BUFENTER
to be applied:

    apply_autocmds_retval(EVENT_BUFENTER, NULL, NULL, FALSE, curbuf, &retval);

Closes #1353