tag.c: Convert mfp to NUL delimited string for all mtt
In 8.0.0190, matching tags were added to a hash table as 0x1 delimited
strings.  After finding all matches, the 0x1 were supposed to be
converted back to NUL so the rest of the tag code, and importantly
parse_match(), didn't see any change in behavior.

However, the loop that converts the bytes back was not skipping over the
initial byte of the tag, which indicates the table in which the tag was
stored.  Since the table corresponding to the ctags table is represented
by a 0 byte, this was causing the loop to exit without performing any
replacements, thus breaking parse_match()'s expectations.