Correct syntax in lstlisting w/out g:tex_verbspell
In my setup of Vim, `verbatim` environments get highlighted just fine.
The `lstlisting` environments do not have any special syntax
highlighting. Characters like `$` or `_` screw up the syntax
highlighting of all the following LaTeX code although they are not
supposed to be interpreted as LaTeX.

From the source it seems that support for `lstlisting` is built-in, but
for some reason only when `g:tex_verbspell` is available and enabled.
The only difference between the two blocks seems to be the
`contains=@Spell`, so I have just copied the `lstlisting` line to the
other branch.