Improve parsing of Urxvt and SGR mouse escape codes.
Urxvt mouse escape codes look like this: ESC[35;123;321M
When ttymouse=urxvt, <UrxvtMouse> would be set to only ESC[, and
then some logic later in check_termcode would react to that, and parse
the rest of the data after it. However, ESC[ is also the prefix of some
other escape codes, like <zHome> (ESC[8~). So, those would stop
working when ttymouse=urxvt, as they are then read as <UrxvtMouse>8~.

This patch changes <UrxvtMouse> to not match just ESC[, but match the
full escape code: ESC[*M, where the asterisk is a wildcard. Such a
'wildcard' functionality was already available for ';*X' and '<M-O>*X',
and is now extended to work for any '*X'. This way, the <UrxvtMouse>
escape code is no longer a prefix of other escape codes. The logic in
check_termcode parsing the mouse commands is changed accordingly.

This patch does the same for <SgrMouse>, for ttymouse=sgr. However, SGR
mouse escape codes for release events end in 'm', instead of 'M'. Now
the full escape code is matched, instead of just the ESC[ prefix, these
different escape codes now have different keys in keymap.h: <SgrMouse>
and <SgrMouseRel>.

Tested in Urxvt v9.22 with ttymouse=urxvt, and in Xterm 327 with
ttymouse=sgr.