screen.c: account for translated string length
`[RO]` is appended to the status line and `len` is increased with the length of
this string (4). However, the string is marked for translation and may thus well
be larger (or smaller) than 4. Therefore, we check the length at runtime. The
resulting len is never actually used, and thus could be removed.  However, by
keeping this line, the body of this if-statement is kept consistent with
surrounding code, and future changes can not forget to add this line when
additional strings are added to p.