Recognize javascript filetype from shebang
Node.js is a JavaScript runtime which is currently quite popular.  It is
typically referenced with `#!/usr/bin/env node` except on Debian-based
distributions where it is referenced as `#!/usr/bin/env nodejs`.

Rhino is a JavaScript engine written in Java, first released in 1998,
and although it is significantly less popular than Node.js for scripting
it is still often used for scripting in Java environments.

Recognize both shebang lines as the javascript filetype.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>