BUG: Pack dirs not yet in 'rtp' when normal plugins run
Some plugins provide autoload functions to customize them, or for reuse
by other plugins.  If that client plugin is in a pack directory itself,
all is fine (because of :h packload-two-steps). However, if the client
is in a "normal" ~/.vim/plugin location, the pack autoload path is not
yet in 'runtimepath'; that is added only _after_ normal plugins have
been run.  The client needs to explicitly
    :runtime START autoload/foo.vim
or the autoload call will fail.

This behavior is bad, and contradicts the documentation under :h
packages, which reads:

> When Vim starts up, after processing your .vimrc, it scans all directories in
> 'packpath' for plugins under the "pack/*/start" directory.  First all those
> directories are added to 'runtimepath'.  Then all the plugins are loaded.

The current behavior actually is:
When Vim starts up, after processing your .vimrc _and normal plugins_,
it scans all directories in 'packpath' for plugins under the
"pack/*/start" directory.  First all those directories are added to
'runtimepath'.  Then all the _pack_ plugins are loaded.

The problem is that in vim_main2(), the :packloadall command (via
ex_packloadall(), which performs both adding and loading) is reused,
after the normal plugins have already been sourced.

To solve the issue, factor add_all_pack_dirs() and load_all_pack() (the
two steps) out of ex_packloadall(), and invoke the former before
sourcing plugins.  This way, 'runtimepath' is already augmented with the
pack plugin paths (but no pack plugin has yet been run), and references
to autoload functions in pack plugins work as expected.