Prevent a buffer-overflow when parsing the qf results
We need to re-allocate a bigger buffer even when the old one has the
same exact size of the text we're going to write over using vim_strncpy.
This is obvious in hindsight as vim_strncpy expects 'out' to be len+1
bytes long, failure to do so resulted in silent memory corruption
leading to a crash.