Regression: Pack plugins are loaded twice (after 8.0.0612)
A side effect of adding the pack dirs earlier to 'runtimepath' (by calling add_pack_start_dirs()) is that this call that was originally meant to only load the traditional plugins now also loads the pack plugins as well (fortunately, in the right order):
    source_runtime((char_u *)"plugin/**/*.vim", DIP_ALL | DIP_NOAFTER)

The following invocation of load_start_packages() is superfluous and must be dropped to avoid duplicate loading of pack plugins.  Instead, only set the did_source_packages flag (as would have done by the dropped call) to neutralize the effect of :packloadall.

The existing Test_after_comes_later() verifies the ordering of plugin loads, but because :scriptnames just records the _first_ load of a script, it cannot ensure that each plugin is loaded only once.
Record each plugin load in g:sequence, and assert ordering (again) and uniqueness (new) in a separate check.