os_unix: treat GNU screen in xterm as xterm
Problem:    The xterm background color detection does not work when Vim
            is running in GNU screen.
Solution:   Treat GNU screen on xterm as xterm -- it is supposed to
            support the same escape sequences. (Lubomir Rintel)

GNU screen recently gained support for OSC 11 sequence that allows us to
detect the terminal background color and pick the right color scheme.
However, termcap does not support the non-standard xterm OSC sequences and they
work only with native xterm support. GNU screen seeks to support the
xterm OSC sequences and in fact indicates that it's running in xterm in
its $TERM variable. Let's trust it that it's xterm compatible in that
case and use the internal xterm terminal support.

Tested on GNU screen git master on GNOME Terminal (that pretends to be
xterm-256color) and xterm.