[configure.ac] Define MACOS_X for all macOS builds
Currently, MACOSX is defined only when the darwin feature is enabled.

The variable has been used to specify platform-specific build parameters
needed for the if_* modules and the Carbon GUI to compile and link.

But this is strange since those modules are inherently irrelevant to the
feature.

Also, there's no good reason to exclude any of the modules from
non-darwin featured builds.

To correct those unreasonable things,

1.  Abandon MACOSX.
2.  Define the new variable MACOS_X for all macOS build, and have it
    take over the role of MACOSX.
3.  There has been a place where MACOSX is truly meant for the darwin
    feature, that is, the block where Cocoa is conditionally added to
    $LIBS.  To this particular usage, it isn't reasonable to replace
    MACOSX with MACOS_X.  To address this, introduce a new variable
    called MACOS_X_UNIX dedicated to the block.

Be careful not to confuse this MACOS_X shell variable with the C macro
of the same name currently defined in vim.h and used throughout the
source files.  That possible confusion will be resolved with some of
patches of this patch set afterwards by changing the current usage of
the C macro meticulously so that its semantics will be consistent with
that of the shell variable.

For macOS builds, making the names of the shell variables and those of
the corresponding C macros consistent is one of the goals of the patch
set.  Note that, with this patch, that goal is partially accomplished
for MACOS_X_UNIX.