Finetune the selection of object files and frameworks.
Currently, when the darwin feature is enabled, Vim is unconditionally
linked against Cocoa, but that's too much for most of the cases.

For Tiny, os_macosx.m is excluded from the build to avoid link errors.
Then, if +multi_byte is not required for the build, we could exclude
os_mac_conv.c, too.  In other words, we don't need the empty
os_mac_conv.o to build an -multi_byte Vim.

As to Small, it's possible for Vim to have +clipboard with +multi_byte.
But that obviously compromises the interoperability of the clipboard.

This patch corrects those unreasonable things.

The following table summarizes the proposed changes:

                          (*1)
Feature             Clipboard   Extra Files             Framework
====================================================================
      -multi_byte   disabled    None                    None
tiny  --------------------------------------------------------------
      +multi_byte   disabled    os_mac_conv.c           CoreServices
--------------------------------------------------------------------
      -multi_byte   enabled     os_macosx.m             AppKit (*2)
                                os_mac_conv.c (#3)
small --------------------------------------------------------------
      +multi_byte   enabled     os_macosx.m             AppKit
                                os_mac_conv.c
--------------------------------------------------------------------

*1: For macOS, FEAT_CLIPBOARD is defined in vim.h for FEAT_SMALL.
*2: The linker automatically adds CoreServices in favor of AppKit.
*3: As mentioned above, +clipboard prefers +multi_byte.