The true goal and overview of this patch set
Make the Mac-related feature list items easier to use by overhauling
macOS-relevant part of the build system.

The result can be skimmed easily only by reading the changes of a few
lines in runtime/doc/eval.txt.

With the changes, non-darwin Vim on macOS returns 1 for has('mac')
instead of 0.

To accomplish that, we could introduce a new compile flag to indicate
that the host system is macOS.  This is the easiest way to address the
issue, but there's a thing to note.

Why has non-darwin Vim returned 0 for has('mac')?  Why has that been
overlooked or remainted unfixed?

Personally, I think one of the reasons is that the transposition of
MACOS_X and MACOS_X_UNIX; the former is defined only if the latter is
defined.  While MACOS_X looks like covering all builds, it actually
doesn't.  It's a shorthand of MACOS_X_UNIX.  Very counter-intuitive.
MACOS_X should represent all the builds, and MACOS_X_UNIX should
represent the subset of MACOSX_X covering the darwin feature; by that,
the subset of non-darwin builds are represented even better.

Also, in src/configure.ac, what is referred to as "darwin" in our
documents is represented by the shell variable MACOSX, while in the
source code, the same notion is represented by MACOS_X_UNIX or MACOS_X
as its shorthand.  Context-dependent various symbols for the same single
notion.  Even if one can handle them with care, they are definitely very
confusing.  Preferably, they should be renamed so that people can work
with them comfortably.

Hence, it is true that adding a new flag solves the current problem
quite easily, but the whole situation will get more convoluted.  We need
an overhaul to get it back to be managable.

To accomplish that, this patch changes the semantics of the C macro
MACOS_X so that it will represent all the builds on macOS.

In theory, for darwin builds, the change shouldn't affect them except
for Tiny without multi_byte since it's identical with that for
non-darwin.

For non-darwin builds, we have to examine each occurance of MACOS_X,
MACOS_X_UNIX and MACOS in the source code one by one to determine what
exactly it has been meant for, and correct it without breaking darwin
builds if necesary.

Basically, patches following this commit are for that laborious task;
they will fix breakages or side effects resulting from the proposed
changes with this commit.

To begin with, they will fix breakages in the build system so that we
will run 'make' and 'make test' successfully for every possible build on
macOS.  After them, patches for fixing breakages or side effects which
are hard to catch will follow.