Fix breakages of Tiny darwin build without multi_byte and its test.
This is the only case where a darwin build falls back to that for
non-darwin (*1).

The cause of the breakages is that os_mac.h is now included even for
non-darwin builds, which results in some conflict with os_unix.h.

This issue could be solved quite easily by not including the header file
against non-darwin builds.

However, since non-darwin builds are those on macOS, people hardly see
why os_mac.h is excluded for them.  Rather, we should modify os_mac.h so
that it can coexist with os_unix.h in favor of the MACOS_X of the new
semantics.

To be more specific,

1.  Change around USE_CR: Without this, 'make test' would abort at a
    very early phase.
2.  Change around MAXPATHL: Without this, MAXPATHL and BASENAMELEN would
    conflict with those coming from os_unix.h.
3.  Change around SIGPROTOARG: Without this, HAVE_AVAIL_MEM would remain
    defined and cause a link error.

With this commit, 'make' and 'make test' should work for all dawrin
builds.

*1: The darwin feature was made on top of the previous build system
    where non-darwin was default.  Hence, although the current default
    builds are darwin, this sort of fallback, that restoring to
    non-default value, happens.