Fix breakage of non-darwin builds
If MACOS_CONVERT is set, some of the object files are compiled to have
references to the functions defined in os_mac_conv.c which implement
part of the darwin feature.

Therefore, the macro shouldn't be defined for non-darwin builds.

With this commit, 'make' and 'make test' should work for all non-darwin
builds except 'make test' for the GUIs on XQuartz.