Differentiate the GUIs on XQuartz from the Carbon GUI in 'guioptions'
Since MACOS_X is now defined for non-darwin builds, MACOS is defined for
them, too.  This makes Test_set_guioptions() fail against the GUIs on
XQuartz.

To fix that, MACOSX is replaced with FEAT_GUI_MAC.  Obviously,
FEAT_GUI_MAC would be still more appropriate than MACOS here even if
the semantics of MACOS_X wasn't changed.  It is worth doing that for
it's own sake.

With this commit, 'make test' should work for all non-darwin builds, in
particular, the GUIs on XQuartz.