Change MACOS_X_UNIX to MACOS_X_DARWIN
After all, MACOS_X_UNIX is meant for taking advantage of macOS-specific
API rather than for making use of Unix-proper API.

Probably, this is the last piece to prevent the lasting misunderstanding
and confusion.

Still, the name remains confusing because Darwin is the name of the
kernal of macOS.  Hopefully, someday someone will give a better name to
what we call "the darwin feature of Vim" today.