Implement an additional test for "gettabvar()"
This test is intended to confirm a bug fix to the "win_remove()"
function in the "window.c" source file.

It is because the "win_remove()" function updates "firstwin" but
does not update "curtab->tp_firstwin" that "f_gettabvar()" in the
"evalfunc.c" source file fails to correctly identify the first
window in the current tab page.  Thus, "gettabvar()" fails when it
shouldn't.