Add assignment operations in "win_remove()"
The "win_remove()" function in "window.c" did not simultaneously set
the "firstwin" and the "curtab->tp_firstwin" pointers.  This meant
that there was a chance of them being momentarily out of sync.

For example, "gettabvar()" checks that the first window in the
referenced tab is valid before getting the requested variable.  If
"gettabvar()" is called from the tabline after the first window is
closed (which calls "win_remove()" indirectly), the "firstwin" and
the "curtab->tp_firstwin" pointers will differ!

The above will cause "gettabvar()" to fail unexpectedly.