Get rid of MACOS_CLASSIC
Although the support for the platform in question has been discontinued
since Vim 7, there're still remnants of it.  Remove them.

N.B. At some places, MAC_OS_CLASSIC has been used.  This may be a typo
or a direct reference to a macro defined in a system header file.
Either way, it is supposed to share the same fate with MACOS_CLASSIC.