Prevent unnecessary XQuartz dependency of non-darwin Tiny and Small
For non-darwin Tiny and Small builds for Mac OS X having XQuartz
installed, the resulting executable is linked against X11 libraries
since mblen() is inadvertently replaced with _Xmblen().

To fix that, add a check for mblen() to configure.ac, add a check for
HAVE_MBLEN to after the checks for __amigaos4__ and __ANDROID__, move
the preprocessor block for replacing mblen() with _Xmblen() to the place
where Xlocale.h is included in order to make the code cleaner, and
correct the indentation level of a #if block to reflect the nesting
depth correctly.

This commit should also fixes unexpected X11 dependency seen in some
multibyte-enabled builds on other platforms.