ftplugin/python: fix completion if compiled only with +python3
In Debian, the default vim version is not compiled with +python, but
with +python3. In this case, pythoncomplete#Complete() will throw an error as it
depends on +python. However, we can use python3complete#Complete() if
+python3 is available.

If no Python support was compiled, pythoncomplete#Complete() will be
used and still throw an error, so the user is warned in this case.

Signed-off-by: Roland Hieber <r.hieber@pengutronix.de>