Parse unified diff
This makes it possible to either parse an "ed" style diff (which is the
traditional diff) or parse the unified diff.

The unified diff is created by the xdiff library. Note: it is highly
experimental. Note: it only works, if there are zero context lines.

Still this allows a 'diffexpr' to return a unified diff and parse it
correctly, as long as it is in a unified diff with zero context lines.