term.c: fix use of wrong variable name
commit 89894aa671ed1db03d95d38 introduced a check that the
prev_winpos_.. variable are set correctly, however it used the variable
prev_winpos_y twice and did not test prev_winpos_x at all.

Most likely it was intended to test prev_winpos_x and prev_winpos_y to
be larger -1 and set the new position then. So fix it to the desired
state.