Add support for ft/syntax bash
Currently, there is no language/filetype/syntax called "bash", it's the
sh syntax/filetype with a buffer-local variable `b:is_bash` that
determines whether or not bash-specific features are enabled when
processing the script.

The problem is that a) `set [ft|syntax]=bash` does not work, and more
importantly, b) as a result, it is not possible to set the filetype to
bash via a modeline, as `#vim: let b:is_bash=1` isn't legal.

This patch will let `set ft=bash` or `set syntax=bash` act as if
`let b:is_bash=1` were defined and then `set ft=sh` or `set syntax=sh`
was called, i.e.:

    set ft=bash
    let b:is_bash "returns 1
    echo &ft "returns sh