Avoid SwapFile Exists warning for Tests
This is not stricktly required for the xdiff feature but an enhancement
to the test suite in general. By default, whenever a SwapFile Exists
message is shown in the Test suite, the Test will stop waiting
indefinitely until a key is pressed. However this will cause an
indefinite hang on the CI infrastructure which will then eventually time
out.

So avoid this at all by globally setting `:set shortmess+=A`. Since one
test however does check the SwapFileExist warning, it needs to be
adjusted to remove the 'A' flag from the shortmess setting.