Add licensing information
The Vim source code is compatible with the GNU GPL so it should be
possible to link in the additional libxdiff library which is distributed
under the GNU LGPL license.

Since the xdiff library is distributed under LGPL 2.1 license, add the
license file to the imported source code. Note, that each file has its
own copyright section which also mentions who is the author of the code.

http://www.xmailserver.org/xdiff-lib.html
https://www.gnu.org/copyleft/lesser.html

Note, that this includes the xdiff library as it is distributed by the
git project (https://git-scm.com/), which is heavily modified from the
upstream project (http://www.xmailserver.org/xdiff-lib.html). So it does
not contain the complete upstream source but only the needed parts for
git (which happen to be also enough for Vim).