Use utf-8 encoding for child Vim in Test_terminal_dumpwrite_composing
The test is relying on the child Vim process to display the composed
characters a certain way in order to test term_dumpwrite().  However, if
the tests aren't running in a locale that defaults the child Vim's
'encoding' to utf-8, then the text will look different than expected
(e.g.  "a~@ e~B o~H").

Forcing the child Vim to use encoding=utf-8, just like the main Vim is
doing, fixes the issue.