Fix "jump-motions" docs to make sure to include all the jump commands
Previously, the docs for "jump-motions" lists a set of commands that will trigger a jump in Vim, but it's not inclusive, nor does it make it clear that operator-pending motions are still counted (e.g. Doing `=G` to indent the file till the end will trigger a jump since "G" itself does).  Fix the docs to include most of the possible operations that will trigger jumps.